[ios] use file-saver to save files rather than trying to do it myself
