+++ app/helpers/download.ts
@@ -1,14 +1,7 @@
+import { saveAs } from 'file-saver';
+
 export function downloadDataURL(filename, data) {
-  const element = document.createElement('a');
-  element.setAttribute('href', data);
-  element.setAttribute('download', filename);
-
-  element.style.display = 'none';
-  document.body.appendChild(element);
-
-  element.click();
-
-  document.body.removeChild(element);
+  saveAs(dataURLtoBlob(data), filename);
 }
 
 export function downloadText(filename, text) {
@@ -16,4 +9,13 @@ export function downloadText(filename, text) {
     filename,
     `data:text/plain;charset=utf-8,${encodeURIComponent(text)}`,
   );
+}
+
+function dataURLtoBlob(dataurl) {
+  var arr = dataurl.split(','), mime = arr[0].match(/:(.*?);/)[1],
+      bstr = atob(arr[1]), n = bstr.length, u8arr = new Uint8Array(n);
+  while(n--){
+      u8arr[n] = bstr.charCodeAt(n);
+  }
+  return new Blob([u8arr], {type:mime});
 }
