+++ package.json
@@ -57,6 +57,7 @@
     "@babel/polyfill": "7.8.7",
     "@babel/runtime": "7.9.2",
     "browser-update": "^3.3.16",
+    "file-saver": "^2.0.2",
     "hammerjs": "2.0.8",
     "heic2any": "^0.0.3",
     "i18next": "^19.4.2",
